Day 9

+++ 5_test.go
@@ -18,8 +18,8 @@ func TestDay5(t *testing.T) {
 	program, err := intcode.ReadIntCode(input)
 	require.NoError(t, err)
 	tc := []VMTC{
-		{Program: program, Input: []int{1}, ExpectedReturnValue: 3, ExpectedOutput: []int{0, 0, 0, 0, 0, 0, 0, 0, 0, 4601506}},
-		{Program: program, Input: []int{5}, ExpectedReturnValue: 314, ExpectedOutput: []int{5525561}},
+		{Program: program, Input: []int64{1}, ExpectedReturnValue: 3, ExpectedOutput: []int64{0, 0, 0, 0, 0, 0, 0, 0, 0, 4601506}},
+		{Program: program, Input: []int64{5}, ExpectedReturnValue: 314, ExpectedOutput: []int64{5525561}},
 	}
 	IntcodeVMTest(t, tc)
 }
